Enable JSON deserialisation of missing fields
Jsonb cannot construct an immutable object unless all of its field
values are known. If a field is missing, the object can only be
deserialised if it's mutable.

We currently have objects that have been persisted where null-value
fields were elided from the JSON representation, so we cannot use
immutability. Either the Jsonb implementation needs to change, or
we would have to use an alternative library like Jackson.