[SR-366] provided default unknown risk info where systems lack risk lens
solves issue in risklist only. Where a system does not have all the risk
lenses associated with it, the risk list does not know what to display.
A default value of "unknown" is now passed to the risk list in this
circumstance. The system still does not have the risk associated with
it.